ОНО РАБОТАЕТ!!!!

+++ Services/WorldOfPowerTools.API/Controllers/UserController.cs
@@ -1,17 +1,29 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using WorldOfPowerTools.API.Extensions;
 using WorldOfPowerTools.Domain.Enums;
 using WorldOfPowerTools.Domain.Repositories;
+using WorldOfPowerTools.Domain.Services;
 
 namespace WorldOfPowerTools.API.Controllers
 {
     [Route("api/[controller]/")]
     [ApiController]
+    [Authorize]
     public class UserController : ControllerBase
     {
+        private readonly Actions GetAllAccess = Actions.Users;
+        private readonly Actions GetByIdAccess = Actions.Users;
+        private readonly Actions RemoveAccess = Actions.Users;
+        private readonly Actions AddUserRightsAccess = Actions.Users;
+        private readonly Actions RemoveUserRightsAccess = Actions.Users;
+
+        private readonly SecurityService _securityService;
         private readonly IUserRepository _userRepository;
 
-        public UserController(IUserRepository userRepository)
+        public UserController(SecurityService securityService, IUserRepository userRepository)
         {
+            _securityService = securityService;
             _userRepository = userRepository;
         }
 
@@ -19,6 +31,8 @@ namespace WorldOfPowerTools.API.Controllers
         [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<IActionResult> GetAll(int skip = 0, int? count = null)
         {
+            if (!_securityService.UserOperationAvailability(User.GetUserRights(), GetAllAccess))
+                return StatusCode(StatusCodes.Status405MethodNotAllowed, "У вас нет доступа к этой операции");
             return Ok(await _userRepository.GetAllAsync(skip, count));
         }
 
@@ -27,6 +41,8 @@ namespace WorldOfPowerTools.API.Controllers
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> GetById(Guid id)
         {
+            if (!_securityService.UserOperationAvailability(User.GetUserRights(), GetByIdAccess))
+                return StatusCode(StatusCodes.Status405MethodNotAllowed, "У вас нет доступа к этой операции");
             return await _userRepository.GetByIdAsync(id) is { } item ? Ok(item) : NotFound();
         }
 
@@ -35,6 +51,8 @@ namespace WorldOfPowerTools.API.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> RemoveUser(Guid id)
         {
+            if (!_securityService.UserOperationAvailability(User.GetUserRights(), RemoveAccess))
+                return StatusCode(StatusCodes.Status405MethodNotAllowed, "У вас нет доступа к этой операции");
             try
             {
                 var userId = await _userRepository.RemoveByIdAsync(id);
@@ -51,6 +69,8 @@ namespace WorldOfPowerTools.API.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> AddUserRights(Guid userId, Actions action)
         {
+            if (!_securityService.UserOperationAvailability(User.GetUserRights(), AddUserRightsAccess))
+                return StatusCode(StatusCodes.Status405MethodNotAllowed, "У вас нет доступа к этой операции");
             try
             {
                 var user = await _userRepository.GetByIdAsync(userId);
@@ -70,12 +90,14 @@ namespace WorldOfPowerTools.API.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> RemoveUserRights(Guid userId, Actions action)
         {
+            if (!_securityService.UserOperationAvailability(User.GetUserRights(), RemoveUserRightsAccess))
+                return StatusCode(StatusCodes.Status405MethodNotAllowed, "У вас нет доступа к этой операции");
             try
             {
                 var user = await _userRepository.GetByIdAsync(userId);
                 if (user == null) throw new Exception("Пользователь по указанному Id не найден");
                 user.ProhibitAction(action);
-                var changedUser = _userRepository.SaveAsync(user);
+                var changedUser = await _userRepository.SaveAsync(user);
                 return Ok(changedUser);
             }
             catch (Exception ex)
